# openapi.yaml
openapi: 3.1.1
info:
  title: Dimiplan API
  description: Dimiplan 서비스를 위한 API 명세서
  version: 1.0.0
  contact:
    name: Dimiplan 개발팀
servers:
  - url: https://dimigo.co.kr:3000
    description: 운영 서버
  - url: http://localhost:8080
    description: 로컬 개발 서버
...
paths:
...
  /auth/google/callback:
    get:
      tags:
        - 인증
      summary: 구글 로그인 콜백
      description: 구글 로그인 후 콜백 처리를 담당합니다.
      parameters:
        - name: code
          in: query
          description: 구글 인증 코드
          required: true
          schema:
            type: string
      responses:
        "302":
          description: 로그인 결과에 따른 리다이렉트
          headers:
            Location:
              schema:
                type: string
              description: 리다이렉트 URL
              examples:
                login_fail:
                  summary: 로그인 실패
                  value: "{FRONT_HOST}/login/fail"
                signup:
                  summary: 사용자 미등록
                  value: "{FRONT_HOST}/signup"
                main:
                  summary: 로그인 완료
                  value: "{FRONT_HOST}"
            Set-Cookie:
              schema:
                type: string
              description: 세션 쿠키 설정 (connect.sid)
              example: connect.sid=s%3A1234abcd.XYZ; Path=/; HttpOnly
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    example: "Authentication successful"
...
  /auth/login:
    post:
      tags:
        - 인증
      summary: 모바일 로그인
      description: 모바일 애플리케이션에서 사용자 ID로 직접 로그인
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  type: string
                  description: 사용자 ID
                email:
                  type: string
                  format: email
                  description: 사용자 이메일
                photo:
                  type: string
                  description: 프로필 이미지 URL
                name:
                  type: string
                  description: 사용자 이름
      responses:
        "200":
          description: 로그인 성공
          headers:
            Set-Cookie:
              schema:
                type: string
              description: 세션 쿠키 설정 (connect.sid)
              example: connect.sid=s%3A1234abcd.XYZ; Path=/; HttpOnly
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                  - sessionId
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  sessionId:
                    type: string
                    example: "s%3A1234abcd.XYZ"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: 서버 오류
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # GitHub Webhook
...
components:
  schemas:
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: 오류 메시지
          example: "Not authenticated"
        error:
          type: string
          description: 상세 오류 정보
          nullable: true
          example: "Error retrieving data: database connection failed"
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: 세션 인증을 위한 쿠키